% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud2xSample.R
\name{cloud2xSample}
\alias{cloud2xSample}
\title{Clip random locations from multiple point cloud datasets (lidar or dap or ifsar etc.)}
\usage{
cloud2xSample(pathClipData = "c:/fusion/clipdata.exe",
  pathCloudmetrics = "c:/fusion/cloudmetrics.exe", pathOutA = "",
  pathOutB = "", pathLasA = "", pathLasB = NA, pathDTMA = NA,
  pathDTMB = NA, patternA = ".*[.]la[s|z]{1}$",
  patternB = ".*[.]la[s|z]{1}$", extentPolyA = NA, extentPolyB = NA,
  samplePoly = NA, proj4A = NA, proj4B = NA, extentSample = c(llx = NA,
  lly = NA, ulx = NA, uly = NA), nSample = 500, radii = list(feet =
  c(FtTenthAc = 37.2, FtAcre = 117.8, Ft5Acres = 263.3), meters = c(MtenthAc =
  11.3, MAcre = 35.9, M5Acres = 80.3))[[1]], sampleShape = c("circle",
  "square")[1], sampleType = c("regular", "random", "hexagonal"),
  doSteps = c("sample", "clip", "metrics"), switchesClipdata = "",
  switchesCloudmetrics = "", procMethod = c("lidR", "FUSION"), nCore = 2,
  temp = "c:\\\\temp\\\\clipdata")
}
\arguments{
\item{pathClipData}{?}

\item{pathCloudmetrics}{?}

\item{pathOutA}{?}

\item{pathOutB}{?}

\item{pathLasA}{?}

\item{pathLasB}{?}

\item{pathDTMA}{?}

\item{pathDTMB}{?}

\item{patternA}{?}

\item{patternB}{?}

\item{extentPolyA}{?}

\item{extentPolyB}{?}

\item{samplePoly}{?}

\item{proj4A}{?}

\item{proj4B}{?}

\item{extentSample}{?}

\item{nSample}{?}

\item{radii}{?}

\item{sampleShape}{?}

\item{sampleType}{?}

\item{doSteps}{?}

\item{switchesClipdata}{?}

\item{switchesCloudmetrics}{?}

\item{procMethod}{?}

\item{nCore}{?}

\item{temp}{?}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2019 May 07 First Debug Version \cr
}
}
\examples{
 <Delete and Replace>

}
\seealso{
\code{\link{another_function}}\cr \code{\link{yet_another_function}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
