% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud2xSample.R
\name{cloudSampleR}
\alias{cloudSampleR}
\alias{cloud2xSample}
\title{Clip random matched locations from multiple point cloud datasets (lidar or dap or ifsar etc.)}
\usage{
cloud2xSample(pathClipData = "c:/fusion/clipdata.exe", pathOutA = "",
  pathOutB = "", pathLasA = "", pathLasB = NA,
  patternA = ".*[.]la[s|z]{1}$", patternB = ".*[.]la[s|z]{1}$",
  pathDTMA = NA, pathDTMB = NA, patternDTMA = ".*[.]asc$",
  patternDTMB = ".*[.]asc$", extentPolyA = NA, extentPolyB = NA,
  sampleShpA = NA, proj4A = NA, proj4B = NA, extentSample = c(llx =
  NA, lly = NA, ulx = NA, uly = NA), nSample = 500, radii = list(feet =
  c(FtTenthAc = 37.2, FtAcre = 117.8, Ft5Acres = 263.3), meters =
  c(MtenthAc = 11.3, MAcre = 35.9, M5Acres = 80.3))[[1]],
  sampleShape = c("circle", "square")[1], sampleType = c("regular",
  "random", "hexagonal"), doSteps = c("sample", "clip", "metrics"),
  switchesClipdata = "", switchesCloudmetrics = "",
  procMethod = c("lidR", "FUSION"), nCore = 2,
  temp = "c:\\\\temp\\\\clipdata")
}
\arguments{
\item{pathClipData}{where to find FUSION clipdata function - only used with procMethod = "FUSION"}

\item{pathOutA}{directory to place clipped plots from project A}

\item{pathOutB}{directory to place clipped plots from project B}

\item{pathLasA}{directory containing project A las or laz files}

\item{pathLasB}{directory containing project B las or laz files}

\item{patternA}{grep pattern to match files on for project A}

\item{patternB}{grep pattern to match files on for project B}

\item{pathDTMA}{(Semi-functional with procMethod = "FUSION", NOT TESTED YET) DTM to use with project A to normalize point heights}

\item{pathDTMB}{(Semi-functional with procMethod = "FUSION", NOT TESTED YET) DTM to use with project B to normalize point heights}

\item{extentPolyA}{polygon of project A extent}

\item{extentPolyB}{polygon of project B extent}

\item{sampleShpA}{(Optional) use an existing shapefile with plot locations}

\item{proj4A}{(Optional) Provide proj4 strings if projects A and B are in different projections but do not contain projection information (e.g. missing .prj files)}

\item{proj4B}{(Optional) Provide proj4 strings if projects A and B are in different projections but do not contain projection information (e.g. missing .prj files)}

\item{extentSample}{(Optional) If the overlap between projects A and B is known, you can provide it here}

\item{nSample}{number of plots to sample}

\item{radii}{a vector of plot radii to use in clipping plots}

\item{sampleShape}{shape of clipped plots ("circle" or "square")}

\item{sampleType}{type of spatial distribution of of sample points ("regular" = square grid,"random","hexagonal" = hexagonal grid)}

\item{doSteps}{(NON FUNCTIONING) Select which steps should be completed in this run}

\item{switchesClipdata}{optional switches to FUSION's cloudmetrics.exe - see fusion documentation for options}

\item{switchesCloudmetrics}{(NON FUNCTIONING)}

\item{procMethod}{Should plots be clipped using R or FUSION?}

\item{nCore}{How many cores should be used for clipping (the number and speed of hdds is as important as the number of machine threads)}

\item{temp}{(use with procMethod = "FUSION")Where should batch files be placed}

\item{pathCloudmetrics}{(NON FUNCTIONING)}
}
\value{
<Delete and Replace>
}
\description{
<Delete and Replace>
}
\details{
<Delete and Replace>

\cr
Revision History
\tabular{ll}{
1.0 \tab 2019 May 07 First Debug Version \cr
1.0 \tab 2019 May 09 Add Documentation for arguments \cr
}
}
\examples{
		library(rgdal)
		library(rgeos)
		library(cloudSampleR)

		#A read in tiles, prepare extents of data
		poly1=rgdal::readOGR("D:/data/wadnr_hood_canal/las/hood_canal_3in_DSM_2015/manage_las","las_polys")
		poly1a=gBuffer(poly1,width = 5)
		poly2=rgdal::readOGR("D:/data/wadnr_hood_canal/las/hood_canal_6in_DSM_2015/manage_las","las_polys")
		poly2a=gBuffer(poly2,width = 5)

		#B clip plots and send to output folders
		cloud2xSample(
			pathClipData = "c:/fusion/clipdata.exe"
			,pathOutA = "d:/temp/hood_canal_test/clip3in/"
			,pathOutB = "d:/temp/hood_canal_test/clip6in/"
			,pathLasA = "D:/data/wadnr_hood_canal/las/hood_canal_3in_DSM_2015/"
			,pathLasB = "D:/data/wadnr_hood_canal/las/hood_canal_6in_DSM_2015/"
			,extentPolyA = poly1a
			,extentPolyB = poly2a
			,nCore = 3
			,nSample = 150
			#,procMethod = "FUSION"
			,procMethod = "lidR"
			,radii = list( feet = c(FtTenthAc = 37.2, FtAcre = 117.8, Ft5Acres = 263.3 ))
		)


}
\seealso{
\code{\link{readLAS}}\cr \code{\link{spsample}}\cr
}
\author{
Jacob Strunk <Jstrunk@fs.fed.us>
}
